requirements.txt
aioredis==1.2.0
asgi-redis==1.4.3
asgiref==3.2.10
astroid==2.0.4
async-timeout==3.0.1
attrs==20.2.0
autobahn==18.10.1
Automat==0.7.0
boto3==1.14.23
botocore==1.17.23
Brotli==1.0.9
certifi==2018.10.15
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
dynaconf==3.1.1
gunicorn==20.0.4
hiredis==1.1.0
hvac==0.10.5
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
Jinja2==2.11.2
jmespath==0.10.0
knox==0.1.0
lazy-object-proxy==1.3.1
loguru==0.5.3
MarkupSafe==1.1.1
msgpack==1.0.0
msgpack-python==0.5.6
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2018.7
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
service-identity==18.1.0
shortuuid==1.0.1
six==1.11.0
sqlparse==0.3.1
twisted==20.3.0
txaio==18.8.1
urllib3==1.25.9
whitenoise==5.1.0
win32-setctime==1.0.2
wincertstore==0.2
wrapt==1.10.11
zope.interface==4.6.0